first commit for the implementation of graphql: code has been reorganized"
