+++ app/helpers/custom_link_renderer.rb
