+++ spec/spec_helper.rb
@@ -5,7 +5,7 @@ require 'rack/test'
 require 'rspec'
 require 'guardian-content'
 require 'pry'
-require_relative '../app/app'
+require_relative '../config/app'
 require_relative '../test/vcr_setup'
 
 module RSpecMixin
